Refactor: Refactor widgets in tutorial screen
